feat(datahub): add CommodityBarsReader/Writer for commodity daily bars storage
Restructure commodity store to follow CQRS pattern with bars subdirectory:
- Add CommodityBarsReader for read operations
- Add CommodityBarsWriter for write operations
- Use year partitioning (market/commodity/bars/YYYY.parquet)
- Add comprehensive unit tests for all operations
