+++ packages/datahub/src/ditto_datahub/stores/market/commodity/__init__.py
@@ -1,10 +1,3 @@
-"""Commodity store module."""
+"""Commodity domain market data stores."""
 
-from ditto_datahub.stores.market.commodity.commodity_reader import (
-    CommodityBarsReader,
-)
-from ditto_datahub.stores.market.commodity.commodity_writer import (
-    CommodityBarsWriter,
-)
-
-__all__ = ["CommodityBarsReader", "CommodityBarsWriter"]
+__all__ = []
