+++ packages/datahub/src/ditto_datahub/stores/market/commodity/bars/bars_reader.py
@@ -1,4 +1,10 @@
-"""Commodity Bars data writer."""
+"""
+Commodity daily bars reader.
+
+Provides read-only access to commodity daily bars data stored in Parquet files
+with year partitioning. Following design document at
+docs/plans/2026-02-09-datahub-cqrs-refactor.md.
+"""
 
 from __future__ import annotations
 
@@ -6,16 +12,14 @@ from pathlib import Path
 
 import polars as pl
 
-from ditto_datahub.models.common import OnDuplicate
-from ditto_datahub.models.storage import WriteStoreResult
 from ditto_datahub.stores.base import ParquetStore, YearlyPartition
 
 
-class CommodityBarsWriter:
+class CommodityBarsReader:
     """
-    Commodity daily bars data writer.
+    Commodity daily bars data reader.
 
-    Provides write access to commodity daily bars data with year partitioning.
+    Provides read-only access to commodity daily bars data with year partitioning.
     Uses composition pattern with ParquetStore for all data operations.
 
     Storage structure:
@@ -34,7 +38,7 @@ class CommodityBarsWriter:
 
     def __init__(self, data_root: Path) -> None:
         """
-        Initialize CommodityBarsWriter.
+        Initialize CommodityBarsReader.
 
         Args:
             data_root: Root directory for data storage.
@@ -42,41 +46,16 @@ class CommodityBarsWriter:
         """
         self._store = ParquetStore(data_root, YearlyPartition())
 
-    # ============ Write operations ============
+    # ============ Read operations ============
 
-    def write(
-        self,
-        df: pl.DataFrame,
-        year: int,
-        on_duplicate: OnDuplicate = OnDuplicate.ERROR,
-    ) -> WriteStoreResult:
-        """
-        Write bars data to the store.
-
-        Args:
-            df: DataFrame to write.
-            year: Year partition.
-            on_duplicate: Duplicate data handling strategy.
-
-        Returns:
-            Write result statistics.
-
-        """
-        return self._store.write(
-            self.DATASET,
-            df,
-            on_duplicate.value,
-            year=year,
-        )
-
-    def delete(
+    def read(
         self,
         instrument_ids: list[int] | None = None,
         start_date: str | None = None,
         end_date: str | None = None,
-    ) -> int:
+    ) -> pl.DataFrame:
         """
-        Delete bars data from the store.
+        Read bars data from the store.
 
         Args:
             instrument_ids: Filter by instrument IDs.
@@ -84,30 +63,52 @@ class CommodityBarsWriter:
             end_date: End date (YYYY-MM-DD).
 
         Returns:
-            Number of deleted records.
+            DataFrame with matching records.
 
         """
-        return self._store.delete(
+        return self._store.read(
             self.DATASET,
             instrument_ids=instrument_ids,
             start_date=start_date,
             end_date=end_date,
         )
 
-    def delete_partition(self, partition_key: str) -> bool:
+    def count(
+        self,
+        instrument_ids: list[int] | None = None,
+        start_date: str | None = None,
+        end_date: str | None = None,
+    ) -> int:
         """
-        Delete a partition by key.
+        Count records in the dataset.
 
         Args:
-            partition_key: Partition key (e.g., "2024").
+            instrument_ids: Filter by instrument IDs.
+            start_date: Start date (YYYY-MM-DD).
+            end_date: End date (YYYY-MM-DD).
 
         Returns:
-            True if deleted, False if file didn't exist.
+            Number of matching records.
 
         """
-        return self._store.delete_partition(self.DATASET, partition_key)
+        return self._store.count(
+            self.DATASET,
+            instrument_ids=instrument_ids,
+            start_date=start_date,
+            end_date=end_date,
+        )
 
     # ============ Metadata operations ============
+
+    def get_years(self) -> list[int]:
+        """
+        Get available years for this dataset.
+
+        Returns:
+            Sorted list of available years.
+
+        """
+        return self._store.get_years(self.DATASET)
 
     def get_checksum(self, partition_key: str) -> str:
         """
@@ -122,6 +123,26 @@ class CommodityBarsWriter:
         """
         return self._store.get_checksum(self.DATASET, partition_key)
 
+    def get_date_range(self) -> tuple[str | None, str | None]:
+        """
+        Get overall date range for the dataset.
+
+        Returns:
+            Tuple of (start_date, end_date) as strings, or (None, None) if empty.
+
+        """
+        return self._store.get_date_range(self.DATASET)
+
+    def list_instrument_ids(self) -> list[int]:
+        """
+        List unique instrument IDs in the dataset.
+
+        Returns:
+            Sorted list of unique instrument IDs.
+
+        """
+        return self._store.list_instrument_ids(self.DATASET)
+
     @property
     def data_root(self) -> Path:
         """
@@ -132,6 +153,3 @@ class CommodityBarsWriter:
 
         """
         return self._store.data_root
-
-
-__all__ = ["CommodityBarsWriter"]
